refactor: App 클래스가 사용자의 상호작용을 담당하도록 getUserPrice 메서드 Price 클래스에서 이동
